+++ CHANGELOG.md
@@ -4,6 +4,7 @@
 Release 2022-01-02
 - Fix error in setting new version (#1)
 - Update donation editing form with UIkit styles, fix missing amount (#2)
+- Limit phpblock to one payment button (#3)
 
 ## Version 0.1.1
 Release 2022-01-02